Refactor controllers and improve chessboard rendering logic
Replaced `MainController` and `BoardController` with a unified `ApiController` to streamline functionality and improve maintainability. Updated `Store` and related services to better handle configuration and state management. Modified HTML, CSS, and rendering logic for improved visuals and fen update handling.
