+++ index.html
@@ -29,12 +29,22 @@
     <script type="module">
       import { Chessfield } from './src/chessfield.ts';
 
-      Chessfield(document.getElementById('board-1'), {});
-      Chessfield(document.getElementById('board-2'), {
-        fen: 'r2q2k1/1p6/p2p4/2pN1rp1/N1Pb2Q1/8/PP1B4/R6K b - - 2 25',
-      });
+      // Chessfield(document.getElementById('board-1'), {});
+      // Chessfield(document.getElementById('board-2'), {
+      //   fen: 'r2q2k1/1p6/p2p4/2pN1rp1/N1Pb2Q1/8/PP1B4/R6K b - - 2 25',
+      // });
 
       const chessfieldTv = Chessfield(document.getElementById('board-3'));
+      setTimeout(() => {
+        console.log('change 1');
+        chessfieldTv.setFen('2pN1rp1/1p6/p2p4/r2q2k1/N1Pb2Q1/8/PP1B4/R6K b - - 2 25');
+        chessfieldTv.setFen('2pN1rp1/1p6/p2p4/r2q2k1/N1Pb2Q1/8/PP1B4/R6K b - - 2 25');
+      }, 3000);
+      setTimeout(() => {
+        console.log('change 2');
+        chessfieldTv.setFen('r2q2k1/1p6/p2p4/2pN1rp1/N1Pb2Q1/8/PP1B4/R6K b - - 2 25');
+        chessfieldTv.setFen('r2q2k1/1p6/p2p4/2pN1rp1/N1Pb2Q1/8/PP1B4/R6K b - - 2 25');
+      }, 10000);
 
       // const stream = fetch('https://lichess.org/api/games/user/neio',{headers:{Accept:'application/x-ndjson'}});
       fetch('https://lichess.org/api/tv/feed')
@@ -65,7 +75,7 @@
                 try {
                   const json = JSON.parse(line);
                   const fen = json.d.fen;
-                  // chessfieldTv.updateFen(fen);
+                  // chessfieldTv.setFen(fen);
                 } catch (e) {
                   console.error(`JSON Parse Error: ${e}`);
                 }
@@ -84,16 +94,16 @@
   </head>
   <body>
     <div>
+      board with a fen updated
+      <div class="chessfield" id="board-3"></div>
+    </div>
+    <div>
       basic board, default config
       <div class="chessfield" id="board-1"></div>
     </div>
     <div>
       board with a fen
       <div class="chessfield" id="board-2"></div>
-    </div>
-    <div>
-      board with a fen updated
-      <div class="chessfield" id="board-3"></div>
     </div>
 
     <script type="module" src="/src/chessfield.ts"></script>
